Modifying the function
In here i'm making the code mpre readable to fixing bugs, also i am adding the exceptions types to the endpoint which will be commited later

+++ Dtos/BasketDto/DishBasketDto.cs
@@ -4,9 +4,9 @@
     {
         public string Name { get; set; } = string.Empty;
 
-        public int Price { get; set; }
+        public double Price { get; set; }
 
-        public int TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
 
         public int Amount { get; set; }
